pyproject.toml
[project]
name = "family-photos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.8",
    "gunicorn>=23.0.0",
    "pillow>=12.0.0",
    "pillow-heif>=1.1.1",
    "python-dotenv>=1.2.1",
    "whitenoise>=6.11.0",
]
